refactor(drafts): DraftZone 抽出 read/edit/publish 状态机 hook
DraftZone.tsx 531 → ~230 行，退化为瘦渲染组件。

- useDraftZone.ts：read/edit/publish 状态机 hook —— 全部 state（isEditing /
  editingBody / saving / publishing / publishError / confirmDelete）+ ref（textarea +
  4 个同步 ref + isMutating 锁）+ effect（ref 同步 / mutate 锁 / unmount cleanup /
  registerEditTrigger / draft.body 同步 / focus / Esc window listener）+ handler
  （runCancelLogic / handleSave / handlePublish / handlePublishFromEdit / handleCancel /
  handleDelete / handleConfirmDelete / onKeyDown）+ 派生（canEdit / canSave / isStash）
- DraftZone.tsx：消费 hook 做渲染（head / edit / read / foot / 错误 / ConfirmModal）

取消四档逻辑（取消按钮 / unmount cleanup / Esc 共用 runCancelLogic 读 ref 最新值）整体
搬进 hook，行为不变；eslint-disable 兜底原样保留。DraftZone 导出名与路径不变。
